gitlab-ci: update and extend used checks
This also reworks flake.nix a bit to pass statix and somewhat deadnix.

This also removes distribution archives so everyone need to use either
Gitlab generated snapshot of repository or git directly. This is
reaction on attack of xz distribution archives. We do not have to ship
them anyway. There is a little to no incentive to do so right now.